permission checking for views

+++ server/api/views/letter.py
@@ -2,7 +2,7 @@ from rest_framework import status
 from rest_framework.decorators import APIView
 from rest_framework.response import Response
 
-from api.models import Letter
+from api.models import Letter, Inbox
 from api.serializers import LetterSerializer
 
 
@@ -11,16 +11,22 @@ class LetterView(APIView):
         letter_id = request.query_params.get('id')
         try:
             letter = Letter.objects.get(pk=letter_id)
-            serializer = LetterSerializer(letter)
-            return Response(serializer.data)
+            if Inbox.objects.filter(user=request.user, letter=letter).exists():
+                serializer = LetterSerializer(letter)
+                return Response(serializer.data)
+            else:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
         except Letter.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
     def post(self, request):
         serializer = LetterSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
+            if Letter.objects.get(pk=serializer.validated_data['id']).sender == request.user:
+                serializer.save()
+                return Response(serializer.data)
+            else:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -28,7 +34,10 @@ class LetterView(APIView):
         try:
             letter_id = request.data.get('id')
             letter = Letter.objects.get(id=letter_id)
-            letter.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if letter.sender == request.user:
+                letter.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
         except Letter.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
